+++ database/migrations/2025_05_22_093243_create_page_multimenu_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('page_multimenu')) {
+            return;
+        }
+
         Schema::create('page_multimenu', function (Blueprint $table) {
             $table->id();
             $table->foreignId('page_id')->constrained()->onDelete('cascade');
